Split array packer tokenizer into header and source files
